requirements.txt
black
flask
numpy
opencv-python-headless
python-dotenv
requests
ultralytics
